+++ typescript-react/src/types/todo.tsx
@@ -1,0 +1,5 @@
+export type Todo = {
+  id: string;
+  title: string;
+  completed: boolean;
+};
